+++ sw.js
@@ -67,7 +67,7 @@ function mapKindToRoute(kind, roomId) {
   }
 }
 
-var CACHE_NAME = 'boost-20260528033737';
+var CACHE_NAME = 'boost-20260529083406';
 var ASSETS = [
   './',
   './index.html',